Refactor: Restructure project hierarchy.
- Rename WebdriverInfo to WebdriverDownloadInfo
- Remove driver_management module
- move driver_management/mod.rs into webdriver_download_info.rs
