+++ webdriver-downloader/src/common/installation_info.rs
