+++ webdriver-downloader/src/common/url_info.rs
