+++ webdriver-downloader/src/common/verification_info.rs
