+++ webdriver-downloader/src/driver_impls/geckodriver_info/os_specific/macos.rs
