+++ webdriver-downloader/src/lib.rs
@@ -1,3 +1,7 @@
-pub use driver_management::*;
+pub use common::*;
+pub use driver_impls::*;
+pub use webdriver_download_info::*;
 
-mod driver_management;
+mod common;
+mod driver_impls;
+mod webdriver_download_info;
