+++ webdriver-downloader/tests/main.rs
@@ -6,13 +6,9 @@ use mockall::mock;
 use reqwest::IntoUrl;
 use tempfile::TempDir;
 
-use webdriver_downloader::common::installation_info::{
-    InstallationError, WebdriverInstallationInfo,
-};
-use webdriver_downloader::common::url_info::{UrlError, WebdriverUrlInfo, WebdriverVersionUrl};
-use webdriver_downloader::common::verification_info::{
-    VerificationError, WebdriverVerificationInfo,
-};
+use webdriver_downloader::installation_info::{InstallationError, WebdriverInstallationInfo};
+use webdriver_downloader::url_info::{UrlError, WebdriverUrlInfo, WebdriverVersionUrl};
+use webdriver_downloader::verification_info::{VerificationError, WebdriverVerificationInfo};
 
 mod driver_management;
